Refactored exceptions into Exceptions.hpp. Wrapped in babs_ecs namespace.
